Issue 303 PR: Recommended migration doc change.
TL;DR  I conclude removal of the Before and After example removes a source of confusion.  
Payload property has been removed from the Action interface.  I really like the "Before" and "After" concept of this migration guide however in this case, the change seems to me to be only observable via intellisense or IDE.  I may well be wrong, but I just don't see how the change is reflected in the Before and After Case provided.  I think do think that users need to understand that when a payload is appropriate it may be added, but that the Action interface will no longer enforce inclusion of a payload property.

In the future I will be more proactive with my pull requests as I hope to be an increasingly active member this and other Repos that I frequent.  (Newbie).  

Many thanks for the great work 
Steve