improve documentation for metaReducers
At the first glance as a new user of ngrx it was difficult to understand whats a metaReducer is! After some researches i found in https://netbasal.com/implementing-a-meta-reducer-in-ngrx-store-4379d7e1020a that a metaReducer is something similar to middleware or rather it can achieve the same effects with metaReducers.