docs(data): improve DefaultDataServiceConfig root parameter docs
add more details for the DefaultDataServiceConfig api docs and Entity Dataservice
doc page.  specifically, added details and examples for providing a remote api to
the DefaultDataServiceConfig root parameter.  Current docs do not include this
information, and it this makes it clear how one would use a remote api, which is
a very common use case.

Closes #2071