feat: add disable forRoot() guard check
In some complex applications, it may be necessary to provide a way to disable the forRoot() guard check. This is very rare, but in some instances it's required when dealing with applications that are sandboxed and built in a parent application without using Angular Elements. By injecting the DISABLE_ROOT_EFFECTS_GUARD or DISABLE_ROOT_STORE_GUARD in a parent injector, this will allow the guard to be bypassed for advanced use cases.