Remove ComparerStr
The compare function is used in two places, neither of which expect it to be able to return a string:
The first caller (https://github.com/ngrx/platform/blob/master/modules/entity/src/sorted_state_adapter.ts#L174) is the Array prototype sort function, and there it "should return a negative, zero, or positive value, depending on the arguments" (https://www.w3schools.com/js/js_array_sort.asp).
The second caller (https://github.com/ngrx/platform/blob/master/modules/entity/src/sorted_state_adapter.ts#L187) does a numerical comparison with the result.

Even though an id can be a string, the result of a comparison shouldn't be.

This will be a breaking change; any client that returned a string from their comparer will have a compilation failure. Note that any implementation that would not compile because of this would not have been working correctly in the first place.