test(store): fix strictImmutability tests passing by default
'NgRx' actions were passing tests by default as no case handled them in
the reducer. Adding cases to handle 'NgRx' actions exposed a futher
error in the tests. 'should not throw for NgRx actions' test was testing
and failing against an Unserializable state. The library allows NgRx
actions where the action is Unserializable so the 'should not throw for
NgRx actions' test was moved from State Serialization to Action
Serialization.