fix(component-store): fix memoization not working when passing selectors directly to select
Make selectors created with `createSelector` usable with `ComponentStore.select`
which looks like it should work, but would previously result in the selector getting recomputed each time.

When select method is called with a single argument, the projector function was being passed directly to `map`,
which invokes its callback with an extra `index` argument. Because the value of index is always different,
the memoized selector would get invalidated even if the actual state remained exactly the same.