fix(component-store): tree-shake hooks check from production bundle
This commit allows for tree-shaking `checkProviderForHooks` from production bundles
since its expression is wrapped with `isDevMode`, which is supposed to be called only
in development mode. This will reduce the number of change detection cycles in production
for existing component stores due to the use of the `asapScheduler`, which schedules a macrotask
within the Angular zone. This leads to redundant change detections in production because the
callback acts as a "noop" (as `isDevMode` returns false).