fix(signals): destroy hook doesn't run in injection context.
The Signal Store saves the injection context upon instantiation and re-uses it in the `onDestroy` hook.
If the Store is provided in `root`, any test with an `onDestroy` hook fails with the error message: `NG0205: Injector has already been destroyed`.

This commit requires `withHook` do access the DI outside of `onDestroy`.