feat(signals) Protection for Overriding Properties
This Proof of Concept intrdocues a feature which prevents
that feature override properties (`state`, `signals`, `methods`).

It does that by adding a conditional type `NoOverride` which
as a constraint to the parameters of `signalState`.

Example:
signalState(
  withState({ id: 1, name: 'hallo', prettyName: 'hi' }),
  withComputed((store) => {
    return {
      prettyName: computed(() => store.name()),
    };
  })
)

This would fail to compile because `prettyName` is overriden.

The overriding protection is enabled by default. It is very likely
that overriding doesn't happen on purpose.

An opt-out is also possible via `signalState({allowOverrides: true})`.