fix(store-devtools): remove `unwrapAction`
`unwrapAction` internally runs `eval` if the argument
is of type `string`.

It is unclear why that `unwrapAction` was necessary
in the first place, and it looks like an artifact
from a older codebase. The line was added
in 2018.

Since `eval` is not best practice and according to #4213,
esbuild has issues with it, it is removed.