perf(signals): avoid creating unnecessary objects in excludeKeys
Make the `excludeKeys` create only one object as it's enough for the function to work correctly.
Currently the `excludeKeys` helper function creates a new object on every iteration of the non-removed keys.
It's unnecessary as - before the function returns the result outside, nothing can have the reference to the
newly created object. It's safe to mutate the temporarily created result.

no breaking changes, fixes #4238