feat(store): enable action dispatching on signal changes
Functions that return actions, with property values derived from signals,
can now also be dispatched.

```typescript
class BookComponent {
  bookId = input.required<number>();
  store = inject(Store);

  constructor() {
    this.store.dispatch(() => loadBook({ id: this.bookId() }));
  }
}
```

`dispatch` executes initially and every time the `bookId` changes.
If `dispatch` is called within an injection context, the signal is
tracked until the context is destroyed.

When `dispatch` is called outside a component's injection context,
the signal is tracked globally throughout the application's lifecycle.

By default, the injection context of the caller is used. If the call happens
outside an injection context, the Store will use its own injection context,
which is usually the root injector.

It is also possible to provide a custom injector as the second parameter:

```typescript
this.store.dispatch(
  () => loadBook({ id: this.bookId() }),
  { injector: this.injector }
);
```